Pulling protected member functions out to public so they can be called in static method.  Fixing a number of compiler warnings from XCode on mac.

+++ vrpn_Analog_Radamec_SPI.C
@@ -549,7 +549,9 @@ void	vrpn_Radamec_SPI::mainloop()
 	    struct timeval current_time;
 	    vrpn_gettimeofday(&current_time, NULL);
 	    if ( duration(current_time,timestamp) > MAX_TIME_INTERVAL) {
-		    sprintf(errmsg,"Timeout... current_time=%ld:%ld, timestamp=%ld:%ld",current_time.tv_sec, current_time.tv_usec, timestamp.tv_sec, timestamp.tv_usec);
+		    sprintf(errmsg,"Timeout... current_time=%ld:%ld, timestamp=%ld:%ld",
+					current_time.tv_sec, static_cast<long>(current_time.tv_usec),
+					timestamp.tv_sec, static_cast<long>(timestamp.tv_usec));
 		    SPI_ERROR(errmsg);
 		    status = STATUS_RESETTING;
 	    }
